Squashed commit of the following:
commit 6f664bc8dea22910ac3a840e87dfd67132963bd2
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Sat Jul 15 21:06:22 2023 +1000

    inv move

commit 906585b46f3879f1e7986ad7f4305a236dd37a34
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Sat Jul 15 16:42:04 2023 +1000

    args that make it crash shouldnt be passed in

commit 9c697dc27fa761134f98f6edbdd0ed99a464b405
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Sat Jul 15 16:41:28 2023 +1000

    formatting

commit 1f7fee697e084570251e570b3fb062c6a4637c9d
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Sat Jul 15 16:40:53 2023 +1000

    e

commit 9507558a3688e67c9dc8158939d44b65fffc772d
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Sat Jul 15 16:40:16 2023 +1000

    0 usage

commit c0249441a20018f67e7e89c0ca578ec9ba3c85aa
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Sat Jul 15 16:35:33 2023 +1000

    mixin and event

commit 4344eaaa9a859ff17d7fe09f33bafb4d92a67dd0
Merge: 14f9dd9 9728433
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Sat Jul 15 13:40:17 2023 +1000

    Merge branch 'nonvg' of https://github.com/K-ov/Arsenic into nonvg

commit 14f9dd972059db8213d28351fddcbc79bc563c53
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Sat Jul 15 13:40:05 2023 +1000

    event game loop

commit 9728433bcda534e8fb3ce8b85f9ab85929b4431c
Merge: 8e25cb2 b0d5f46
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Sat Jul 15 13:35:49 2023 +1000

    Merge branch 'Cosmic-SC:nonvg' into nonvg

commit b0d5f46932ebf987e6c3b2ca8f3acd78c3786759
Author: Cosmic <quantumo7@outlook.com>
Date:   Fri Jul 14 15:17:31 2023 +0400

    forgor to add to ModuleManager ahhhhhhhhhhhhhhhh

commit 21fa46c916bb690ad9bf43d85427fb573f1f5ed7
Author: Cosmic <quantumo7@outlook.com>
Date:   Fri Jul 14 15:13:33 2023 +0400

    cure InvMove from cancer + NoHurtCam

commit 8e25cb2dc087025500060bf3549d5c25309a3d35
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Fri Jul 14 16:59:19 2023 +1000

    sprint

commit b3391df73d9a081a601adabe42898c36ee839ecd
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Fri Jul 14 16:56:25 2023 +1000

    aim assist

commit 638b94541bafd65cae6e198c9590be9d7a9bafec
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Fri Jul 14 16:52:49 2023 +1000

    keep everything seperate...

commit 7a13ad79307afca15a8da62cee5e5a0791d7634d
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Fri Jul 14 16:52:29 2023 +1000

    if it doesnt work theres no point

commit 0b265c0eea04a89b2774f2a50036c5ad4a2f6a22
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Fri Jul 14 16:44:51 2023 +1000

    remove bloat

commit d3f6630b45c58efbe0eef9557be49265b07d4e51
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Fri Jul 14 16:44:35 2023 +1000

    redo mixin

commit 3c3cc39c28f436bb49a291d143a7ff2e92cb518f
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Fri Jul 14 16:39:58 2023 +1000

    gui

commit 4c0170ecd1c54ab17df28c38309a357b17fc720b
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Fri Jul 14 16:31:55 2023 +1000

    removed forge event

commit a618bcbab7db50a41299a62fd0148c362752258e
Author: K-v <58765519+K-ov@users.noreply.github.com>
Date:   Fri Jul 14 16:27:44 2023 +1000

    put stuff back

commit 1084b2dc11d2a1e504c1c310f447fbda29c635d4
Author: Cosmic <quantumo7@outlook.com>
Date:   Thu Jul 13 18:05:32 2023 +0400

    Aura Rotations & Clicker improvements, WTap and some misc stuff

commit 7f60fbf7fa04642bcce1e03574c14c1f06187416
Author: Cosmic <quantumo7@outlook.com>
Date:   Thu Jul 13 14:22:15 2023 +0400

    forgor to make this yooooooooo

commit f568eead6d593ee5e596bcaea1414ea0ee7e56e0
Author: Cosmic <133380776+Cosmic-SC@users.noreply.github.com>
Date:   Thu Jul 13 14:13:09 2023 +0400

    Delete .idea directory

commit efccd1c120512491eb7c288a4c899a22245db0b9
Author: Cosmic <quantumo7@outlook.com>
Date:   Thu Jul 13 14:12:03 2023 +0400

    git ignore $$$

commit 0c921dbeebd210fa99001cad2935107716b55404
Author: Cosmic <quantumo7@outlook.com>
Date:   Thu Jul 13 14:04:48 2023 +0400

    aaaaaaaaaaaaaaa

+++ src/main/java/arsenic/event/impl/EventJump.java
@@ -5,9 +5,9 @@ import arsenic.event.types.Event;
 
 public class EventJump extends CancellableEvent implements Event {
     private float yaw;
-    private double motion;
+    private float motion;
 
-    public EventJump(float yaw, double motion){
+    public EventJump(float yaw, float motion){
         this.yaw = yaw;
         this.motion = motion;
     }
@@ -20,11 +20,11 @@ public class EventJump extends CancellableEvent implements Event {
         this.yaw = yaw;
     }
 
-    public double getMotion(){
+    public float getMotion(){
         return this.motion;
     }
 
-    public void setMotion(double motion){
+    public void setMotion(float motion){
         this.motion = motion;
     }
 }
